+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * Added `scannerM` in `Text.MMark.Extension` and `runScannerM` in
   `Text.MMark`.
 
+* Added support for block quotes.
+
+* Added support for unordered and ordered lists.
+
 ## MMark 0.0.1.1
 
 * Fixed a bug in skipping of headers (only one newline after the header line